requirements.txt
streamlit 
pandas 
python-docx 
reportlab 
openpyxl
psycopg2-binary